@@ -4,13 +4,19 @@
     "type": "symfony-bundle",
     "license": "MIT",
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.1",
         "ext-json": "*",
         "ext-intl": "*",
-        "symfony/framework-bundle": "^4.0 || ^5 || ^6",
-        "symfony/translation": "^4.0 || ^5 || ^6",
-        "symfony/security-bundle": "^4.0 || ^5 || ^6",
+        "symfony/framework-bundle": "^5 || ^6",
+        "symfony/translation": "^5 || ^6",
+        "symfony/security-bundle": "^5 || ^6",
         "twig/twig": "^2.0 || ^3"
+    },
+    "require-dev": {
+        "friendsofphp/php-cs-fixer": "^3"
+    },
+    "scripts": {
+        "cs": "./vendor/bin/php-cs-fixer fix"
     },
     "autoload": {
         "psr-4": {
