@@ -7,22 +7,20 @@ namespace Arthem\Bundle\LocaleBundle\Listener;
 use Arthem\Bundle\LocaleBundle\LocaleResolver;
 use Arthem\Bundle\LocaleBundle\Model\UserLocaleInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpFoundation\Session\SessionInterface;
+use Symfony\Component\HttpFoundation\Exception\SessionNotFoundException;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Security\Http\Event\InteractiveLoginEvent;
 use Symfony\Component\Security\Http\SecurityEvents;
 
 class UserLocaleSubscriber implements EventSubscriberInterface
 {
-    private SessionInterface $session;
-    private LocaleResolver $localeResolver;
-
-    public function __construct(SessionInterface $session, LocaleResolver $localeResolver)
-    {
-        $this->session = $session;
-        $this->localeResolver = $localeResolver;
+    public function __construct(
+        private readonly RequestStack $requestStack,
+        private readonly LocaleResolver $localeResolver,
+    ) {
     }
 
-    public function onInteractiveLogin(InteractiveLoginEvent $event)
+    public function onInteractiveLogin(InteractiveLoginEvent $event): void
     {
         $user = $event->getAuthenticationToken()->getUser();
         if (!$user instanceof UserLocaleInterface) {
@@ -30,7 +28,10 @@ class UserLocaleSubscriber implements EventSubscriberInterface
         }
 
         if (null !== $user->getLocale() && $user->getLocale()) {
-            $this->session->set('_locale', $this->localeResolver->getClosestLanguage($user->getLocale()));
+            try {
+                $this->requestStack->getSession()->set('_locale', $this->localeResolver->getClosestLanguage($user->getLocale()));
+            } catch (SessionNotFoundException $e) {
+            }
         }
     }
 
